Move slack commands and routes to common
